+++ ci/script.sh
@@ -1,36 +1,11 @@
 #!/usr/bin/env bash
 
+NAME=${TRAVIS_REPO_SLUG#*/}
+
 function main {
-    # Check if any command failed
-    ERROR=false
-
-    GOARCH=${_GOARCH}
-    GOOS=${_GOOS}
-
-    if [[ ! ${GOARCH} ]]; then
-        exit
-    fi
-
-    env GOOS=${GOOS} GOARCH=${GOARCH} GOARM=${GOARM} go build -o ${NAME} || ERROR=true
-
-    mkdir -p dist
-
-    if [[ ${GOARCH} == "arm64" ]]; then
-        FILE=${NAME}_${TRAVIS_BRANCH}_${GOOS}_arm8
-    else
-        FILE=${NAME}_${TRAVIS_BRANCH}_${GOOS}_${GOARCH}${GOARM}
-    fi
-
-    tar -czf dist/${FILE}.tgz ${NAME} || ERROR=true
-
-    if [[ ${GOOS} == "linux" && ${GOARCH} == "amd64" ]]; then
-        make all || ERROR=true
-        rm dist/gotop
-    fi
-
-    if [ ${ERROR} == "true" ]; then
-        exit 1
-    fi
+    rustup target add $TARGET
+    cargo build --target $TARGET --release
+    tar -czf $NAME-$TRAVIS_TAG-$TARGET.tar.gz -C ./target/$TARGET/release/ $NAME
 }
 
 main